move summable_array to utils
